make snapshot tests explicit in each component
closes #196
