+++ components/x-styling-demo/__tests__/x-styling-demo.test.jsx
@@ -1,0 +1,6 @@
+const snapshots = require('@financial-times/x-test-utils/snapshots');
+const stories = require('../stories');
+
+describe('x-styling-demo', () => {
+	snapshots(stories);
+});
