+++ components/x-teaser/__tests__/x-teaser.test.jsx
@@ -1,0 +1,6 @@
+const snapshots = require('@financial-times/x-test-utils/snapshots');
+const stories = require('../stories');
+
+describe('x-teaser', () => {
+	snapshots(stories);
+});
