+++ package.json
@@ -22,8 +22,8 @@
     "eslint-plugin-react": "^7.10.0",
     "espree": "^4.1.0",
     "fetch-mock": "^6.5.2",
-    "react": "^16.3.1",
-    "react-test-renderer": "^16.3.1"
+    "jest": "^24.7.1",
+    "react": "^16.3.1"
   },
   "x-dash": {
     "engine": {
@@ -38,8 +38,5 @@
     "components/*",
     "packages/*",
     "tools/*"
-  ],
-  "devDependencies": {
-    "jest": "^24.7.1"
-  }
+  ]
 }
