+++ packages/x-test-utils/package.json
@@ -11,10 +11,12 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@financial-times/x-engine": "file:../x-engine",
     "enzyme": "^3.6.0",
     "enzyme-adapter-react-16": "^1.5.0",
     "jest-enzyme": "^6.0.4",
     "react": "^16.5.0",
-    "react-dom": "^16.5.0"
+    "react-dom": "^16.5.0",
+    "react-test-renderer": "^16.8.6"
   }
 }
